BACKUPS/agents_nl2sql_20250905_130916/requirements.txt
langchain>=0.2.10
langgraph>=0.2.22
pydantic>=2.7
python-dotenv>=1.0
rich>=13.7
